feat: i18n support for export
