+++ main/helpers/constants.ts
@@ -15,6 +15,7 @@ export const WD14 = "WD14";
 export const GPTV = "GPTV";
 export const CAPTION = "CAPTION";
 export const GPT_VISION_OPTIONS = "GPT_VISION_OPTIONS";
+export const LOCALE = "LOCALE";
 export const FEEDBACK = "FEEDBACK";
 export const DATASET = "DATASET";
 export const DATASETS = "DATASETS";
