+++ renderer/next-i18next.config.js
@@ -17,6 +17,8 @@ module.exports = {
       "pl", // Polish
     ],
   },
+  debug: process.env.NODE_ENV === "development",
+  reloadOnPrerender: process.env.NODE_ENV === "development",
   localePath:
     typeof window === "undefined"
       ? require("node:path").resolve("./public/locales")
