+++ renderer/next.config.js
@@ -1,6 +1,5 @@
 const transpileModules = require("next-transpile-modules");
 const withTM = transpileModules(["@mui/joy"]); // Pass the modules you would like to see transpiled
-const i18next = require("./next-i18next.config.js");
 
 // Define an array of routes that need rewrites and redirects
 
@@ -13,7 +12,6 @@ const nextConfig = {
   images: {
     unoptimized: true,
   },
-  i18n: i18next.i18n,
 };
 /**
  *
