+++ renderer/pages/[locale]/marketplace.tsx
@@ -1,18 +1,16 @@
-import React from "react";
 import Head from "next/head";
 import { Box, Sheet, Stack, Typography } from "@mui/joy";
-import { Layout } from "@/organisms/layout";
-import { serverSideTranslations } from "next-i18next/serverSideTranslations";
-import { GetStaticPropsContext, InferGetStaticPropsType } from "next";
+import { InferGetStaticPropsType } from "next";
 import { useTranslation } from "next-i18next";
 import { Lottie } from "@/organisms/lottie";
+import { getStaticPaths, makeStaticProps } from "@/ions/i18n/getStatic";
 
 export default function Page(
   _properties: InferGetStaticPropsType<typeof getStaticProps>,
 ) {
   const { t } = useTranslation(["common"]);
   return (
-    <Layout>
+    <>
       <Head>
         <title>{`Captain | ${t("common:marketplace")}`}</title>
       </Head>
@@ -49,14 +47,9 @@ export default function Page(
           </Box>
         </Box>
       </Stack>
-    </Layout>
+    </>
   );
 }
 
-export async function getStaticProps(context: GetStaticPropsContext) {
-  return {
-    props: {
-      ...(await serverSideTranslations(context.locale ?? "en", ["common"])),
-    },
-  };
-}
+const getStaticProps = makeStaticProps(["common"]);
+export { getStaticPaths, getStaticProps };
