+++ renderer/public/locales/es/common.json
@@ -41,9 +41,11 @@
   "openFolder": "Abrir Carpeta",
   "pages": {
     "dataset": {
+      "chooseCaptioningMethod": "\"Elija un método de subtitulación",
       "customCaptionsWithGPTVision": "Subtítulos Personalizados con GPT-Vision",
       "generateCaptionWithBLIP": "Generar Subtítulos con BLIP",
       "generateTagsWithWD14": "Generar Etiquetas con WD14",
+      "oneTimeDownloadNote": "Nuestros métodos avanzados de subtitulación requieren una descarga única de los modelos necesarios, lo que puede llevar unos momentos.",
       "proceedWithGPTVision": "Continuar con GPT-Vision",
       "selectImage": "Seleccionar Imagen",
       "warningOpenAI": "Atención: Usar GPT-Vision enviará tus imágenes a OpenAI. Por favor, cumple con los <1>Términos de Uso</1> de OpenAI."
